World: restructured directory
